requirements.txt
azure-functions
azure-identity
azure-storage-blob
requests
sendgrid
